refactor: mapper 파일 store 디렉토리로 이동 및 movieMapper로 파일명 변경
